Fixed bug 17401, crash on i386 Linux due to off_t/size_t usage, switched to uint64_t instead

+++ src/mediascan.c
@@ -1164,7 +1164,7 @@ void ms_scan_file(MediaScan *s, const char *full_path, enum media_type type) {
   int ret;
   uint32_t hash;
   int mtime = 0;
-  size_t size = 0;
+  uint64_t size = 0;
   DBT key, data;
   char tmp_full_path[MAX_PATH];
